Refactor List.jsx component
