+++ src/Presentation/SpaWeb/src/components/Items/List/Header/Header.css
@@ -57,9 +57,3 @@
   border-radius: 2px;
   height: 2.66667rem;
 }
-
-.default-item-search {
-  border: 1px solid #e3e6ef;
-  background: #fff;
-  padding: 2rem;
-}
