+++ src/Presentation/SpaWeb/src/components/Items/List/List.jsx
@@ -1,12 +1,12 @@
 import React, { Fragment, useState, useEffect } from "react";
-import { Container, Row, Col, Dropdown } from "react-bootstrap";
+import { Container, Row, Col } from "react-bootstrap";
 import "react-input-range/lib/css/index.css";
-import "./List.css";
 import useItemsSearch from "../../../utils/hooks/useItemsSearch";
 import useDebounce from "../../../utils/hooks/useDebounce";
 import { ItemsContainer } from "./ItemsContainer";
-import { Search } from "./Search";
+import { Search } from "./Search/Search";
 import { useParams } from "react-router-dom";
+import { Header } from "./Header/Header";
 
 export const List = () => {
   const [state, setState] = useState({
@@ -44,27 +44,7 @@ export const List = () => {
       <Container>
         <Row>
           <Col lg={12}>
-            <div className="generic-header">
-              <div>
-                <h4>All Items</h4>
-                <p>Total Listing Found: {totalItemsCount}</p>
-              </div>
-              <div className="generic-header-toolbar" role="toolbar">
-                <Dropdown>
-                  <Dropdown.Toggle
-                    className="sort-btn"
-                    variant="light"
-                    id="dropdown-basic"
-                  >
-                    sort by
-                  </Dropdown.Toggle>
-
-                  <Dropdown.Menu>
-                    <Dropdown.Item href="#/action-1">TODO</Dropdown.Item>
-                  </Dropdown.Menu>
-                </Dropdown>
-              </div>
-            </div>
+            <Header totalItemsCount={totalItemsCount} />
           </Col>
           <Col lg={12}>
             <Row>
