+++ src/Presentation/SpaWeb/src/components/Items/List/Search/Search.css
@@ -1,0 +1,5 @@
+.default-item-search {
+  border: 1px solid #e3e6ef;
+  background: #fff;
+  padding: 2rem;
+}
